Exr: clean plugins by adding EXRWriterDefinitions.hpp
Split EXRDefinitons.
