+++ plugins/image/io/Exr/src/reader/EXRReaderPlugin.hpp
@@ -65,4 +65,3 @@ protected:
 }
 
 #endif
-
