+++ plugins/image/io/Exr/src/writer/EXRWriterDefinitions.hpp
@@ -1,0 +1,29 @@
+#ifndef EXRWRITER_DEFINITIONS_HPP
+#define EXRWRITER_DEFINITIONS_HPP
+
+#include <EXRDefinitions.hpp>
+
+#include <tuttle/plugin/global.hpp>
+#include <tuttle/plugin/context/WriterDefinition.hpp>
+
+namespace tuttle {
+namespace plugin {
+namespace exr {
+namespace writer {
+
+static const std::string kParamStorageType     = "storage";
+static const std::string kParamStorageScanLine = "scanLine";
+static const std::string kParamStorageTiles    = "tiles";
+
+enum EParamStorage
+{
+	eParamStorageScanLine = 0,
+	eParamStorageTiles
+};
+
+}
+}
+}
+}
+
+#endif
